requirements.txt
# These are all requirements for development
# These requirements are not needed for using the module

# BASE
bitarray~=2.3.0

# ENCODE
attrs~=21.2.0

# TEST
flake8
coverage
mypy
pytest
pytest-cov

# DEPLOY
twine
wheel

# DOCS
sphinx

libais~=0.17
configparser~=5.2.0
lxml~=4.6.3
setuptools~=58.0.4
pyais~=2.1.4